Revert "Added guards to ensure progressive sequence numbers on command topics"
This reverts commit 225103103ac795045c0fdb4a0ed872500733502c.

The guards make it impossible to send commands to the subscriber after
iiwa_hw has been restarted. For some reason, restarting SmartServo
doesn't help either. The only solution is to restart the cabinet which
is a pain in the ass. So let's remove these checks for now.

Thanks to Denis for helping debugging this.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>